RavenDB-24346 feat (document schema) - restructure components into partials, extract `useDocumentSchema` hook, and temporarily disable Document Schema Playground.
