+++ src/Raven.Studio/typescript/common/shell/menu/items/settings.ts
@@ -21,7 +21,6 @@ import Integrations = require("components/pages/database/settings/integrations/I
 import UnusedDatabaseIds = require("components/pages/database/settings/unusedDatabaseIds/UnusedDatabaseIds");
 import RevisionsBinCleaner = require("components/pages/database/settings/revisionsBinCleaner/RevisionsBinCleaner");
 import DocumentSchema = require("components/pages/database/settings/documentSchema/DocumentSchema");
-import DocumentSchemaPlayground = require("components/pages/database/settings/documentSchema/DocumentSchemaPlayground");
 
 export = getSettingsMenuItem;
 
@@ -169,15 +168,16 @@ function getSettingsMenuItem(appUrls: computedAppUrls) {
             css: 'icon-document',
             dynamicHash: appUrls.documentSchema,
         }),
-        new leafMenuItem({
-            route: 'databases/settings/documentSchema/playground',
-            moduleId: reactUtils.bridgeToReact(DocumentSchemaPlayground.default, "nonShardedView"),
-            title: 'Document Schema Playground',
-            nav: false,
-            css: 'icon-document',
-            dynamicHash: appUrls.documentSchemaPlayground,
-            itemRouteToHighlight: "databases/settings/documentSchema",
-        }),
+        // TODO: For now schema playground is not available. https://issues.hibernatingrhinos.com/issue/RavenDB-22142/Schema-Validation
+        // new leafMenuItem({
+        //     route: 'databases/settings/documentSchema/playground',
+        //     moduleId: reactUtils.bridgeToReact(DocumentSchemaPlayground.default, "nonShardedView"),
+        //     title: 'Document Schema Playground',
+        //     nav: false,
+        //     css: 'icon-document',
+        //     dynamicHash: appUrls.documentSchemaPlayground,
+        //     itemRouteToHighlight: "databases/settings/documentSchema",
+        // }),
         new leafMenuItem({
             route: 'databases/settings/revisions',
             moduleId: reactUtils.bridgeToReact(DocumentRevisions.default, "nonShardedView"),
