+++ src/Raven.Studio/typescript/components/pages/database/settings/documentSchema/partials/DocumentSchemaDeleteModal.tsx
